Partially revert "Update default settings for RHEL 5.3"
On reflecting about this further I don't think it's a good idea to
change either the defaults in the script nor the example lines in the
config because some users may be depending on those details.

This partially reverts commit 706b12b. I've left in some minor comment
changes and whitespace fixes from the original commit.

Signed-off-by: Wincent Colaiuta <win@wincent.com>